% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D4-other_helpers.R
\name{data_for_Y.FUN}
\alias{data_for_Y.FUN}
\title{Make the dataset that will be handled by \code{Y.FUN} in \code{PImain()} and \code{PIsens()}}
\usage{
data_for_Y.FUN(data, targeted, Z.form, C.form)
}
\arguments{
\item{data}{A data frame that has been processed by function \code{declare_data()}.}

\item{targeted}{If \code{TRUE}, the principal score model and the Y1 models are fit to inverse propensity score weighted treatment arm data; and the Y0 models are fit to control arm data that are inverse probability of control weighted and principal score weighted. If \code{FALSE}, such weighting is not used.}

\item{Z.form}{A character string specifying the formula for the propensity score model.}

\item{C.form}{A character string specifying the formula for the principal score model.}
}
\description{
Make the dataset that will be handled by \code{Y.FUN} in \code{PImain()} and \code{PIsens()}
}
