% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A-declare_data.R
\name{declare_data}
\alias{declare_data}
\title{Declare data to the package}
\usage{
declare_data(
  data,
  z.var,
  c.var,
  y.var,
  x.vars,
  samp.wt = NULL,
  id = NULL,
  other.vars = NULL
)
}
\arguments{
\item{data}{The input dataset.}

\item{z.var}{Name of the assigned treatment variable.}

\item{c.var}{Name of the compliance type variable.}

\item{y.var}{Name of the outcome variable.}

\item{x.vars}{A character vector containing names of the covariates.}

\item{samp.wt}{Name of sampling weight variable. Default is \code{NULL}.}

\item{id}{Name of id variable. Default is \code{NULL}.}

\item{other.vars}{A vector of names of other variables to be kept in the dataset. Default is \code{NULL}.}
}
\value{
A data frame with just analysis variables, where assigned treatment is renamed \code{z}, compliance type is renamed \code{c}, outcome variable is renamed \code{y}, sampling weight (if used) is renamed \code{s.wt}, and id variable (if used) is renames \code{id}. Covariates (named in \code{x.vars}) and other variables (named in \code{other.vars}) are retained and not renamed.
}
\description{
This function processes the data frame with user-provided variable names into a data frame that is used by the functions that implement the main analysis and sensitivity analysis methods.
}
