% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F2-jobsII_depress3.R
\name{depress3_mus}
\alias{depress3_mus}
\alias{depress3_mus_pimu1}
\alias{depress3_mus_pimu2}
\alias{depress3_mus_emu1}
\alias{depress3_mus_emu2}
\alias{depress3_mus_MS1}
\alias{depress3_mus_MS2}
\alias{depress3_mus_IF1}
\alias{depress3_mus_IF2}
\alias{depress3_mus_IFH1}
\alias{depress3_mus_IFH2}
\alias{depress3_mus.quants_pimu1}
\alias{depress3_mus.quants_pimu2}
\alias{depress3_mus.quants_emu1}
\alias{depress3_mus.quants_emu2}
\alias{depress3_mus.quants_MS1}
\alias{depress3_mus.quants_MS2}
\alias{depress3_mus.quants_IF1}
\alias{depress3_mus.quants_IF2}
\alias{depress3_mus.quants_IFH1}
\alias{depress3_mus.quants_IFH2}
\title{JOBSII-specific: estimate \eqn{\mu} functions (and optionally, quantiles) for the outcome \code{depress3}}
\usage{
depress3_mus(data, estimator, targeted, quants = FALSE, num.p.points = 100)

depress3_mus_pimu1(data)

depress3_mus_pimu2(data)

depress3_mus_emu1(data)

depress3_mus_emu2(data)

depress3_mus_MS1(data)

depress3_mus_MS2(data)

depress3_mus_IF1(data)

depress3_mus_IF2(data)

depress3_mus_IFH1(data)

depress3_mus_IFH2(data)

depress3_mus.quants_pimu1(data)

depress3_mus.quants_pimu2(data)

depress3_mus.quants_emu1(data)

depress3_mus.quants_emu2(data)

depress3_mus.quants_MS1(data)

depress3_mus.quants_MS2(data)

depress3_mus.quants_IF1(data)

depress3_mus.quants_IF2(data)

depress3_mus.quants_IFH1(data)

depress3_mus.quants_IFH2(data)
}
\arguments{
\item{data}{A data frame that has been processed by \code{declare_data()} and --- if needed for targeted \eqn{\mu} estimation --- has been enhanced by (1) having inverse probability of assigned treatment weight incorporated (ie multiplied) in variable \code{s.wt} and (2) adding variables \code{c.wt} and \code{n.wt} (holding principal score weights, ie probabilities of being a complier and of being a noncomplier, respectively). The reason this input is in this form is that this function will be incorporated in the internal function that estimates nuisance parameters, where this internal function (if required for targeted \eqn{\mu} estimation) will estimate these weights and put the data in this form for \eqn{\mu} functions estimation.}

\item{estimator}{One of the estimators listed in the paper: "epi", "pimu" or "emu" (for the simple estimator \eqn{e\pi}, \eqn{\pi\mu} or \eqn{e\mu}), or "MS", "IF" or "IFH" (for the multi-step, IF-as-estimating-function, or Hajek-ized estimator).}

\item{targeted}{If \code{TRUE}, the principal score model and the Y1 models are fit to inverse propensity score weighted treatment arm data; and the Y0 models are fit to control arm data that are inverse probability of control weighted and principal score weighted. If \code{FALSE}, such weighting is not used.}

\item{quants}{If TRUE, compute quantiles of the estimated conditional \eqn{Y|X,Z=0} distribution. Default is FALSE.}

\item{num.p.points}{Number of quantile points. Default is 100.}
}
\value{
If \code{quants} is FALSE, a data frame with variable \code{id} and the mu functions. If \code{quants} is TRUE, a list including this data frame (named \code{mus}) and a matrix named \code{quants} containing p-quantiles for \eqn{X} values in each row of \code{data}.
}
\description{
Customized estimation of \eqn{\mu} functions and conditional \eqn{Y|X,Z=0} quantiles for the outcome \code{depress3} in the paper data example \code{jobsII}
}
\details{
Important note: This function uses the \code{s.wt} variable in \code{data} in the estimation. This is required of any custom function for \eqn{\mu} estimation, even if this variable is all 1 (i.e., no sampling weights), because this variable will be used by the bootstrap to incorporate bootstrap weights.
}
\keyword{internal}
