% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D3-boot_procedures.R
\name{.ahat_from_jacks}
\alias{.ahat_from_jacks}
\title{(internal) Compute acceleration parameter based on jackknife estimates}
\usage{
.ahat_from_jacks(jacks)
}
\arguments{
\item{jacks}{The set of all jackknife estimates.}
}
\value{
The acceleration parameter.
}
\description{
This function is called by function \code{.get_ahat_for_BCa()}. Its purpose is to avoid having to retype the unintuitive formula.
}
\keyword{internal}
