% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C5-beta_quantile.R
\name{.assemble_beta.quant}
\alias{.assemble_beta.quant}
\title{(internal) Assemble OR-based sensitivity analysis for a specified estimator based on main analysis estimates and estimated nuisance functions.}
\usage{
.assemble_beta.quant(
  estimator,
  main,
  nuis,
  quants,
  sens.range = c(-0.1, 0.1),
  sens.step
)
}
\arguments{
\item{estimator}{One of the estimators listed in the paper: "epi", "pimu" or "emu" (for the simple estimator \eqn{e\pi}, \eqn{\pi\mu} or \eqn{e\mu}), or "MS", "IF" or "IFH" (for the multi-step, IF-as-estimating-function, or Hajek-ized estimator).}

\item{main}{The main analysis result from \code{.assemble_main()}}

\item{nuis}{A dataframe with variables \code{s.wt}; grouping variables \code{z} (treatment vs control arm), \code{c} (=1 if complier in treatment arm, 0 otherwise), \code{n} (=1 if noncomplier in treatment arm, 0 otherwise); outcome \code{y}; plus variables holding the requested nuisance functions: \code{z.wt} (for inverse probability of assigned treatment weight), \code{c.wt} and \code{n.wt} (for the two principal score weights), \code{mu.1c}, \code{mu.1n}, \code{mu.0c}, \code{mu.0n} (for the four outcome mean functions).}

\item{quants}{A list of two matrices of conditional outcome quantiles, for estimation of \code{mu.0c} and \code{mu.0n} if \code{sens.type} is "beta.quant".}

\item{sens.range}{A numeric vector of length 2, holding the minimum and maximum values for the sensitivity parameter. If not specified, defaults to (1/3, 3) for \code{sens.type} being OR, GOR or MR; (-1, 1) for \code{sens.type} being MD; and (-.05, .05) for \code{sens.type} being "beta.quant".}

\item{sens.step}{The width of the step by which to increase the sensitivity parameter. If not specified, defaults to 0.05 for mean-based sensitivity analyses and 0.05 for distribution-based sensitivity analyses.}
}
\value{
A matrix with one row (named \code{rho}) holding the values of the sensitivity parameter and two rows for the corresponding CATE and NACE point estimates.
}
\description{
(internal) Assemble OR-based sensitivity analysis for a specified estimator based on main analysis estimates and estimated nuisance functions.
}
\keyword{internal}
