% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B-main.R
\name{.assemble_main}
\alias{.assemble_main}
\title{(internal) Assemble PI-based main analysis for a specified estimator based on estimated nuisance functions.}
\usage{
.assemble_main(estimator, nuis)
}
\arguments{
\item{estimator}{One of the estimators listed in the paper: "epi", "pimu" or "emu" (for the simple estimator \eqn{e\pi}, \eqn{\pi\mu} or \eqn{e\mu}), or "MS", "IF" or "IFH" (for the multi-step, IF-as-estimating-function, or Hajek-ized estimator).}

\item{nuis}{A dataframe with variables \code{s.wt}; grouping variables \code{z} (treatment vs control arm), \code{c} (=1 if complier in treatment arm, 0 otherwise), \code{n} (=1 if noncomplier in treatment arm, 0 otherwise); outcome \code{y}; plus variables holding the requested nuisance functions: \code{z.wt} (for inverse probability of assigned treatment weight), \code{c.wt} and \code{n.wt} (for the two principal score weights), \code{mu.1c}, \code{mu.1n}, \code{mu.0c}, \code{mu.0n} (for the four outcome mean functions).}
}
\value{
A vector of estimates with elements named \code{cace}, \code{nace}, \code{tau.1c}, \code{tau.1n}, \code{tau.0c}, \code{tau.0n}.
}
\description{
(internal) Assemble PI-based main analysis for a specified estimator based on estimated nuisance functions.
}
\keyword{internal}
