% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3-SMDe.R
\name{.compute_lambda_upsilon_SMDe}
\alias{.compute_lambda_upsilon_SMDe}
\title{(internal) Compute SMDe-based lambda or upsilon functions for a value of the sentivity parameter.}
\usage{
.compute_lambda_upsilon_SMDe(nuis, h, get.upsilon = FALSE)
}
\arguments{
\item{nuis}{A dataframe with variables \code{s.wt}; grouping variables \code{z} (treatment vs control arm), \code{c} (=1 if complier in treatment arm, 0 otherwise), \code{n} (=1 if noncomplier in treatment arm, 0 otherwise); outcome \code{y}; plus variables holding the requested nuisance functions: \code{z.wt} (for inverse probability of assigned treatment weight), \code{c.wt} and \code{n.wt} (for the two principal score weights), \code{mu.1c}, \code{mu.1n}, \code{mu.0c}, \code{mu.0n} (for the four outcome mean functions).}

\item{h}{A value of the sensitivity parameter}
}
\description{
This function is called by \code{.assemble_SMDe()}.
}
\keyword{internal}
