% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C5-beta_quantile.R
\name{.compute_mu.0c_beta.quant}
\alias{.compute_mu.0c_beta.quant}
\title{(internal) Compute beta-based mu.0c and mu.0n functions for a value of the sentivity parameter.}
\usage{
.compute_mu.0c_beta.quant(nuis, quants, s.kappa)
}
\arguments{
\item{nuis}{A dataframe with variables \code{s.wt}; grouping variables \code{z} (treatment vs control arm), \code{c} (=1 if complier in treatment arm, 0 otherwise), \code{n} (=1 if noncomplier in treatment arm, 0 otherwise); outcome \code{y}; plus variables holding the requested nuisance functions: \code{z.wt} (for inverse probability of assigned treatment weight), \code{c.wt} and \code{n.wt} (for the two principal score weights), \code{mu.1c}, \code{mu.1n}, \code{mu.0c}, \code{mu.0n} (for the four outcome mean functions).}

\item{quants}{A list of two matrices of conditional outcome quantiles, for estimation of \code{mu.0c} and \code{mu.0n} if \code{sens.type} is "beta.quant".}

\item{s.kappa}{A value of the sensitivity parameter signed-\eqn{\kappa}.}
}
\description{
This function is called by \code{.assemble_beta.quant()}.
}
\keyword{internal}
