% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C1-GOR.R
\name{.compute_mu_epsilon_GOR}
\alias{.compute_mu_epsilon_GOR}
\title{(internal) Compute OR-based mu.0c and mu.0n functions for a value of the sentivity parameter.}
\usage{
.compute_mu_epsilon_GOR(nuis, r, Y.bounds = NULL, get.epsilon = FALSE)
}
\arguments{
\item{nuis}{A dataframe with variables \code{s.wt}; grouping variables \code{z} (treatment vs control arm), \code{c} (=1 if complier in treatment arm, 0 otherwise), \code{n} (=1 if noncomplier in treatment arm, 0 otherwise); outcome \code{y}; plus variables holding the requested nuisance functions: \code{z.wt} (for inverse probability of assigned treatment weight), \code{c.wt} and \code{n.wt} (for the two principal score weights), \code{mu.1c}, \code{mu.1n}, \code{mu.0c}, \code{mu.0n} (for the four outcome mean functions).}

\item{r}{A value of the sensitivity parameter}

\item{Y.bounds}{A two element vector for the lower and upper bounds of the outcome. Needed when modeling a bounded outcome by a quasibinomial model and/or when conducting a GOR-based sensitivity analysis.}
}
\description{
This function is called by \code{.assemble_OR_mu()}.
}
\keyword{internal}
