% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D4-other_helpers.R
\name{.df.rename}
\alias{.df.rename}
\title{Rename variables in a data frame}
\usage{
.df.rename(x, from, to)
}
\arguments{
\item{x}{The data frame.}

\item{from}{Names of the variables to be renamed.}

\item{to}{New names to give the variables.}
}
\description{
Rename variables in a data frame
}
\keyword{internal}
