% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B-main.R
\name{.do_all_main}
\alias{.do_all_main}
\title{(internal) Compute confidence intervals and bias-correct point estimates for a main analysis}
\usage{
.do_all_main(
  pointFUN,
  data,
  boot.num = 999,
  boot.seed = 12345,
  boot.cont.wt = TRUE,
  double.boot = TRUE,
  BCa.CI = TRUE
)
}
\arguments{
\item{pointFUN}{A function constructed to compute the point estimate, with all options wrapped in, so the only argument of this function is \code{data}. This function should output a numeric vector with two estimates, named "cace" and "nace".}

\item{data}{A data frame that has been processed by function \code{declare_data()}.}

\item{boot.num}{Number of bootstrap samples. Default is 999.}

\item{boot.seed}{Seed for reproducibility. Default is 12345.}

\item{boot.cont.wt}{If \code{TRUE} (default), implement continuous weights bootstrap (with weights based on the Dirichlet distribution). If \code{FALSE}, implement classic bootstrap (with weights based on the multinomial distribution). If}

\item{double.boot}{Whether to implement bias correction of point estimate by double bootstrap. If FALSE, implements bias correction of point estimate by single bootstrap only; this substantially reduces computational burden for the beta-quantile sensitivity analysis. Default is FALSE.}

\item{BCa.CI}{Whether to compute BCa confidence intervals. If FALSE, compute percentile intervals only, which is not recommended. However, BCa CIs are computationally burdensome for the beta-quantile sensitivity analysis. Default is TRUE.}
}
\value{
A table with two rows for CACE and NACE, respectively. The table contains columns for the point estimate, the bias-corrected point estimate, the percentile CI, the BCa CI, and also the bootstrap and double bootstrap means of the estimates.
}
\description{
(internal) Compute confidence intervals and bias-correct point estimates for a main analysis
}
\keyword{internal}
