% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-sens.R
\name{.do_all_sens}
\alias{.do_all_sens}
\title{(internal) Compute confidence intervals and bias-correct point estimates for a sensitivity analysis}
\usage{
.do_all_sens(
  pointFUN,
  data,
  boot.num,
  boot.seed = 12345,
  boot.cont.wt = TRUE,
  double.boot = TRUE,
  BCa.CI = TRUE
)
}
\arguments{
\item{pointFUN}{A function constructed to compute the point estimate vector (one estimate for each value of the sensitivity parameter), with all options wrapped in, so the only argument of this function is \code{data}. This pointFUN function should output a two-column matrix with CACE and NACE in the first and second columns, respectively.}

\item{data}{A data frame that has been processed by function \code{declare_data()}.}

\item{boot.num}{Number of bootstrap samples. Default is 999.}

\item{boot.seed}{Seed for reproducibility. Default is 12345.}

\item{boot.cont.wt}{If \code{TRUE} (default), implement continuous weights bootstrap (with weights based on the Dirichlet distribution). If \code{FALSE}, implement classic bootstrap (with weights based on the multinomial distribution). If}

\item{double.boot}{Whether to implement bias correction of point estimate by double bootstrap. If FALSE, implements bias correction of point estimate by single bootstrap only; this substantially reduces computational burden for the beta-quantile sensitivity analysis. Default is FALSE.}

\item{BCa.CI}{Whether to compute BCa confidence intervals. If FALSE, compute percentile intervals only, which is not recommended. However, BCa CIs are computationally burdensome for the beta-quantile sensitivity analysis. Default is TRUE.}

\item{long}{If TRUE, print progress messages to screen. Default is FALSE.}
}
\value{
A list of two matrices for CACE and NACE, respectively. Each matrix contains columns for the point estimate, the bias-corrected point estimate, the percentile CI, the BCa CI, and also the bootstrap and double bootstrap means of the estimates.
}
\description{
(internal) Compute confidence intervals and bias-correct point estimates for a sensitivity analysis
}
\keyword{internal}
