% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D2-nuisance_estimation.R
\name{.estimate_nuisance}
\alias{.estimate_nuisance}
\title{(internal) Estimate nuisance functions needed for effect estimation}
\usage{
.estimate_nuisance(
  data,
  estimator,
  targeted,
  get.sigma = FALSE,
  Z.form = NULL,
  C.form = NULL,
  Y1c.form = NULL,
  Y1n.form = NULL,
  Y0.form = NULL,
  Y.family = NULL,
  Y.bounds = NULL,
  Y.FUN = NULL
)
}
\arguments{
\item{data}{A data frame that has been processed by function \code{declare_data()}.}

\item{estimator}{One of the estimators listed in the paper: "epi", "pimu" or "emu" (for the simple estimator \eqn{e\pi}, \eqn{\pi\mu} or \eqn{e\mu}), or "MS", "IF" or "IFH" (for the multi-step, IF-as-estimating-function, or Hajek-ized estimator).}

\item{targeted}{If \code{TRUE}, the principal score model and the Y1 models are fit to inverse propensity score weighted treatment arm data; and the Y0 models are fit to control arm data that are inverse probability of control weighted and principal score weighted. If \code{FALSE}, such weighting is not used.}

\item{Z.form}{A character string specifying the formula for the propensity score model.}

\item{C.form}{A character string specifying the formula for the principal score model.}

\item{Y1c.form}{Formula for Y|X,Z=1,C=1 model, processed by \code{.clean_nuis_inputs()}.}

\item{Y1n.form}{Formula for Y|X,Z=1,C=0 model, processed by \code{.clean_nuis_inputs()}.}

\item{Y0.form}{Formula for Y|X,Z=0 model, processed by \code{.clean_nuis_inputs()}.}

\item{Y.family}{The family to be used for outcome models, processed by \code{.clean_nuis_inputs()}.}

\item{Y.bounds}{A two element vector for the lower and upper bounds of the outcome. Needed when modeling a bounded outcome by a quasibinomial model and/or when conducting a GOR-based sensitivity analysis.}

\item{Y.FUN}{Optional function for customized estimation of outcome mean functions. For an example, see function \code{jobsII_earn3_mus()} in this package, and see how it is used in the vignette. If use \code{Y.FUN}, arguments for outcome model forms and family do not need to be specified, and are ignored.}
}
\value{
A dataframe with variables \code{s.wt}; grouping variables \code{z} (treatment vs control arm), \code{c} (=1 if complier in treatment arm, 0 otherwise), \code{n} (=1 if noncomplier in treatment arm, 0 otherwise); outcome \code{y}; plus variables holding the requested nuisance functions: \code{z.wt} (for inverse probability of assigned treatment weight), \code{c.wt} and \code{n.wt} (for the two principal score weights), \code{mu.1c}, \code{mu.1n}, \code{mu.0c}, \code{mu.0n} (for the four outcome mean functions).
}
\description{
(internal) Estimate nuisance functions needed for effect estimation
}
\keyword{TODO:}
\keyword{edit}
\keyword{internal}
\keyword{returns}
