% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D5-geom_split_violin.R
\docType{data}
\name{.GeomSplitViolin}
\alias{.GeomSplitViolin}
\title{(internal) ggproto class for split violin plot}
\format{
An object of class \code{GeomSplitViolin} (inherits from \code{GeomViolin}, \code{Geom}, \code{ggproto}, \code{gg}) of length 2.
}
\usage{
.GeomSplitViolin
}
\description{
(internal) ggproto class for split violin plot
}
\keyword{internal}
