% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D3-boot_procedures.R
\name{.get_ahat_for_BCa}
\alias{.get_ahat_for_BCa}
\title{(internal) Estimate acceleration parameter for construction of BCa intervals}
\usage{
.get_ahat_for_BCa(FUN, data)
}
\arguments{
\item{FUN}{The function that will act on \code{data} to implement the estimator. The output of \code{FUN} should be in the form of a scalar, a vector or a matrix, NOT a list.}

\item{data}{A data frame to be used by \code{FUN}. If sampling weights are used, they should come in a variable named \code{s.wt}.}
}
\value{
Estimated acceleration params in the same format as the output of \code{FUN}, i.e., one ahat for each parameter estimate from \code{FUN}.
}
\description{
(internal) Estimate acceleration parameter for construction of BCa intervals
}
\keyword{internal}
