% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D3-boot_procedures.R
\name{.get_BCa.interval}
\alias{.get_BCa.interval}
\title{(internal) Compute a BCa interval based on a point estimate, a set of boot estimates and an acceleration parameter}
\usage{
.get_BCa.interval(theta.boot, theta.hat, alpha = 0.05, ahat)
}
\arguments{
\item{theta.boot}{A vector of boot estimates.}

\item{theta.hat}{The point estimate.}

\item{alpha}{\code{(1-alpha)} is the confidence level.}

\item{ahat}{The acceleration parameter.}
}
\value{
The BCa interval.
}
\description{
(internal) Compute a BCa interval based on a point estimate, a set of boot estimates and an acceleration parameter
}
\keyword{internal}
