% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D3-boot_procedures.R
\name{.get_boot_estimates}
\alias{.get_boot_estimates}
\title{(internal) Compute bootstrap estimates}
\usage{
.get_boot_estimates(
  FUN,
  data,
  boot.num = 999,
  boot.cont.wt = TRUE,
  double.boot = FALSE,
  boot.seed = 12345
)
}
\arguments{
\item{FUN}{The function that will act on \code{data} to implement the estimator. The output of \code{FUN} should be in the form of a scalar, a vector or a matrix, NOT a list.}

\item{data}{A data frame to be used by \code{FUN}. If sampling weights are used, they should come in a variable named \code{s.wt}.}

\item{boot.num}{Number of bootstrap samples. Default is 999.}

\item{boot.cont.wt}{If \code{TRUE} (default), implement continuous weights bootstrap (based on the Dirichlet distribution). If \code{FALSE}, implement the resampling bootstrap.}

\item{double.boot}{If \code{FALSE} (default), implement the standard bootstrap. If \code{TRUE}, implement a single double-bootstrap. This is used for correction of finite sample bias.}

\item{boot.seed}{Seed for reproducibility. Default is 12345.}
}
\value{
Bootstrap estimates in the structure of a vector if \code{FUN} outputs a scalar, a matrix if \code{FUN} outputs a vector, a 3D array if \code{FUN} outputs a matrix. In the last two cases, the last dimension of the array is the dimension of the bootstrap samples.
}
\description{
Implement bootstrap to compute bootstrap estimates.
}
\keyword{internal}
