% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D3-boot_procedures.R
\name{.get_jack_estimates}
\alias{.get_jack_estimates}
\title{(internal) Obtain jackknife estimates for an estimator}
\usage{
.get_jack_estimates(FUN, data)
}
\arguments{
\item{FUN}{The function that will act on \code{data} to implement the estimator. The output of \code{FUN} should be in the form of a scalar, a vector or a matrix, NOT a list.}

\item{data}{A data frame to be used by \code{FUN}. If sampling weights are used, they should come in a variable named \code{s.wt}.}
}
\value{
Jackknife estimates in the structure of a vector if \code{FUN} outputs a scalar, a matrix if \code{FUN} outputs a vector, a 3D array if \code{fun} outputs a matrix. In the last two cases, the last dimension of the array is the dimension of the bootstrap samples.
}
\description{
Compute jackknife estimates.
}
\keyword{internal}
