% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D4-other_helpers.R
\name{.get_mu_hat}
\alias{.get_mu_hat}
\title{Compute mu.hat based on outcome model inputs}
\usage{
.get_mu_hat(form, data, wts, y.fam, pred.dat, get.dispersion = FALSE)
}
\arguments{
\item{form}{Model formula}

\item{data}{Data for model fitting.}

\item{wts}{A vector of weights for model fitting.}

\item{y.fam}{Model family for \code{glm()}.}

\item{pred.dat}{Data to predict mu.hat on}

\item{get.dispersion}{Whether to also compute the dispersion parameter based on the quasi family.}
}
\description{
Compute mu.hat based on outcome model inputs
}
\keyword{internal}
