% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D4-other_helpers.R
\name{.get_rho.vec}
\alias{.get_rho.vec}
\title{Compute the vector of rho values to use for the sensitivity analysis, for a pair of min and max values}
\usage{
.get_rho.vec(sens.range, sens.step)
}
\arguments{
\item{sens.range}{A numeric vector of length 2, holding the minimum and maximum values for the sensitivity parameter. If not specified, defaults to (1/3, 3) for \code{sens.type} being OR, GOR or MR; (-1, 1) for \code{sens.type} being MD; and (-.05, .05) for \code{sens.type} being "beta.quant".}

\item{sens.step}{The width of the step by which to increase the sensitivity parameter. If not specified, defaults to 0.05 for mean-based sensitivity analyses and 0.05 for distribution-based sensitivity analyses.}
}
\description{
Compute the vector of rho values to use for the sensitivity analysis, for a pair of min and max values
}
\keyword{internal}
