% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D4-other_helpers.R
\name{.is.01}
\alias{.is.01}
\title{Check if a vector is binary and coded as 0/1}
\usage{
.is.01(x)
}
\arguments{
\item{x}{A vector}
}
\description{
Check if a vector is binary and coded as 0/1
}
\keyword{internal}
