% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A-declare_data.R
\name{.is.declared}
\alias{.is.declared}
\title{Check if data frame has been processed by function \code{declare_data()}}
\usage{
.is.declared(data)
}
\arguments{
\item{data}{A data frame}
}
\description{
Check if data frame has been processed by function \code{declare_data()}
}
\keyword{internal}
