% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E2-plot_multiple_estimators.R
\name{.plot_sens_multiple}
\alias{.plot_sens_multiple}
\title{(internal) Uber plot of the wholee bunch of estimators}
\usage{
.plot_sens_multiple(sens, to.plot)
}
\arguments{
\item{sens}{A named list of results from different sensitivity analyses from running \code{PIsens()}. The names can be \code{epi1}, \code{epi2}, \code{pimu1}, \code{pimu2}, \code{emu1}, \code{emu2}, \code{MS1}, \code{MS2}, \code{IF1}, \code{IF2}, \code{IFH1}, \code{IFH2} (where the letter part indicates the estimator, and the number part indicates whether the estimator is based on (1) untargeted or (2) targeted nuisance estimation).}

\item{to.plot}{The parameter for which results are to be plotted. Options are: "cace", "nace", "tau.0c", "tau.0n".}
}
\description{
(internal) Uber plot of the wholee bunch of estimators
}
\keyword{internal}
