% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D4-other_helpers.R
\name{.replace_y}
\alias{.replace_y}
\title{(internal) Replace the response variable in a model formula}
\usage{
.replace_y(formula, new.y)
}
\arguments{
\item{formula}{A model formula}

\item{new.y}{Name of new response variable}
}
\description{
(internal) Replace the response variable in a model formula
}
\keyword{internal}
