% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D4-other_helpers.R
\name{.trans_bounds}
\alias{.trans_bounds}
\title{Transformation of a bounded variable (vector) to a different pair of bounds by shifting and scaling}
\usage{
.trans_bounds(x, from, to)
}
\arguments{
\item{x}{The vector to be transformed}

\item{from}{The pair of bounds of \code{x}}

\item{to}{The destination pair of bounds}
}
\description{
Transformation of a bounded variable (vector) to a different pair of bounds by shifting and scaling
}
\keyword{internal}
