% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D4-other_helpers.R
\name{.wtd_mean}
\alias{.wtd_mean}
\title{Weighted mean function from the stats package}
\usage{
.wtd_mean(x, w)
}
\arguments{
\item{x}{Vector of numeric values to be averaged}

\item{w}{Vector of weights}
}
\description{
Weighted mean function from the stats package
}
\keyword{internal}
