% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F1-jobsII_earn3.R
\name{earn3_FUN1}
\alias{earn3_FUN1}
\alias{earn3_FUN1_IFH}
\title{JOBSII-specific: estimate \eqn{\mu} functions for the outcome \code{earn3}}
\usage{
earn3_FUN1(data, targeted, include.mu.1c)

earn3_FUN1_IFH(data)
}
\arguments{
\item{data}{A data frame that is produced within the calling function (\code{PImain()} or \code{PIsens}). See \code{Y.FUN} in documentation of \code{PImain()} or \code{PIsens()} for description of this type of dataset.}

\item{targeted}{If \code{TRUE}, the principal score model and the Y1 models are fit to inverse propensity score weighted treatment arm data; and the Y0 models are fit to control arm data that are inverse probability of control weighted and principal score weighted. If \code{FALSE}, such weighting is not used.}

\item{include.mu.1c}{If FALSE, only estimate mu.0(X), not mu.1c(X) and mu.1n(X).}
}
\value{
A data frame with variable \code{id} and the mu functions.
}
\description{
Customized estimation of \eqn{\mu} functions for the outcome \code{earn3} in the paper data example \code{jobsII}. \code{earn2_FUN1()} is the mother function, from which we derive child functions, eg \code{earn3_FUN1_IFH()}, which are fed into the \code{Y.FUN} argument of \code{PImain()} or \code{PIsens()}.
}
\details{
Important note: This function satisfies what is required of argument \code{Y.FUN} in function \code{PIsens()} uses the \code{s.wt} variable in \code{data} in the estimation. This is required of any custom function for \eqn{\mu} estimation, even if this variable is all 1 (i.e., no sampling weights), because this variable will be used by the bootstrap to incorporate bootstrap weights.
}
