% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B1-main_pieces.R
\name{estimate_weights}
\alias{estimate_weights}
\title{Estimate weights}
\usage{
estimate_weights(data, Z.form = NULL, C.form = NULL, targeted = FALSE)
}
\arguments{
\item{data}{A data frame that has been processed by function \code{declare_data()}.}

\item{Z.form}{A character string specifying the formula for the propensity score model.}

\item{C.form}{A character string specifying the formula for the principal score model.}

\item{targeted}{If TRUE, principal score model is fit to weighted data that target the principal stratum covariate space. If FALSE (default), this weighting is not used. NOTE: if TRUE, \code{Z.form} must be specified.}
}
\description{
Estimate inverse propensity score weights (\code{z.wt}) or principal score weights (\code{c.wt} and \code{n.wt}), or both
}
