% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIsens-package.R
\docType{package}
\name{PIsens-package}
\alias{PIsens-package}
\alias{_PACKAGE}
\title{PIsens: Sensitivity Analyses for Violation of Principal Ignorability in Estimating (Non)Complier Average Causal Effects}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/trangnguyen74/PIsens}
  \item Report bugs at \url{https://github.com/trangnguyen74/PIsens/issues}
}

}
\author{
\strong{Maintainer}: Trang Nguyen \email{trang.nguyen@jhu.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
