% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E1-plot_results.R
\name{plot_finite.sample.bias}
\alias{plot_finite.sample.bias}
\title{Function that makes figure 5 in the paper}
\usage{
plot_finite.sample.bias(
  sens,
  sens.type,
  include.bias.correction = FALSE,
  include.CI = FALSE,
  effect.ylims = NULL,
  outcome.ylims = NULL,
  sens.para.lims = NULL
)
}
\arguments{
\item{sens}{The output of \code{PIsens()}.}

\item{include.bias.correction}{If TRUE, include the curves for bias-corrected point estimates based on the bootstrap and based on the double bootstrap. Default is FALSE.}

\item{include.CI}{If TRUE, include the curves for the point-wise 95\% BCa confidence intervals (for a sense of the scale of the finite sample bias). Default is FALSE.}

\item{effect.ylims}{An optional two-element vector for the y-axis limits of the average effect panel.}

\item{outcome.ylims}{An optional two-element vector for the y-axis limits of the mean Y0 panel.}

\item{sens.para.lims}{An optional two-element vector for the range of the sensitivity parameter to be plotted. Use this if want to limit the range more than is available in \code{sens}.}
}
\description{
Function that makes figure 5 in the paper
}
