% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E1-plot_results.R
\name{plot_sens}
\alias{plot_sens}
\title{Plot the result of a sensitivity analysis}
\usage{
plot_sens(
  sens,
  CI = "BCa",
  point = "corrected1",
  effect.ylims = NULL,
  outcome.ylims = NULL,
  sens.para.lims = NULL
)
}
\arguments{
\item{sens}{Result from function \code{PIsens}.}

\item{CI}{Type of confidence interval to be plotted. Options are "BCa" and "percentile". If want "BCa", \code{sens} must have been the result of running \code{PIsens()} with \code{BCa.CI = TRUE} (which is the default of \code{PIsens()}).}

\item{point}{Type of point estimates to be plotted. Options are "uncorrected", "corrected1" (meaning bias-corrected based on the bootstrap) and "corrected2 (meaning bias-corrected based on the double bootstrap). If want "corrected2", \code{sens} must have been the result of running \code{PIsens()} with \code{double.boot = TRUE} (which is the default of \code{PIsens()}).}

\item{effect.ylims}{Optional. A two-element numeric vector holding the plot limits for the NACE and CACE.}

\item{outcome.ylims}{Optional. A two-element numeric vector holding the plot limits for the potential outcome means.}

\item{sens.para.lims}{Optional. A two-element numeric vector holding the plot limits for the sensitivity parameter. To be used to make publishable plots if analysis was run on a larger range than should be shown, e.g., if the relevant range is narrower.}
}
\description{
Plot the result of a sensitivity analysis
}
