% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1-view_assumptions.R
\name{visualize_GOR}
\alias{visualize_GOR}
\title{\loadmathjax
Figure 2 in the paper}
\usage{
visualize_GOR(point.mu.0 = 0.5, point.c.wt = 0.25)
}
\arguments{
\item{point.mu.0, point.c.wt}{A pair of \mjeqn{\mu_0(X),\pi_c(X)}{ASCII representation} values to be shown as the example point in the plot}
}
\description{
Plot the connection between \mjeqn{\mu_{00}(X)}{ASCII representation} and \mjeqn{\mu_{01}(X)}{ASCII representation} implied by a GOR assumption
}
