% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1-view_assumptions.R
\name{visualize_kappa}
\alias{visualize_kappa}
\title{Make the kappa plot in the paper}
\usage{
visualize_kappa(pi.x = c(0.15, 0.3, 0.5))
}
\arguments{
\item{pi.x}{A vector of values for \eqn{\pi_c(X)} to be plotted.}
}
\description{
Make the kappa plot in the paper
}
