% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1-view_assumptions.R
\name{visualize_MR_implication}
\alias{visualize_MR_implication}
\title{\loadmathjax
Plot the implication of sensitivity MR values (violin plot)}
\usage{
visualize_MR_implication(
  MRs,
  MRlabs,
  data,
  estimator,
  targeted,
  Z.form,
  C.form,
  Y.form,
  Y.model,
  Y.bounds,
  Y.FUN = NULL
)
}
\arguments{
\item{MRs}{Sensitivity MR values to be used for the plot.}

\item{MRlabs}{Labels for MRs to put on the plot. For example, for an MR value of 1/3, one may want to use the label "1/3".}

\item{data}{A data frame that has been processed by function \code{declare_data()}.}

\item{estimator}{One of the estimators listed in the paper: "epi", "pimu" or "emu" (for the simple estimator \eqn{e\pi}, \eqn{\pi\mu} or \eqn{e\mu}), or "MS", "IF" or "IFH" (for the multi-step, IF-as-estimating-function, or Hajek-ized estimator).}

\item{targeted}{If \code{TRUE}, the principal score model and the Y1 models are fit to inverse propensity score weighted treatment arm data; and the Y0 models are fit to control arm data that are inverse probability of control weighted and principal score weighted. If \code{FALSE}, such weighting is not used.}

\item{Z.form}{A character string specifying the formula for the propensity score model.}

\item{C.form}{A character string specifying the formula for the principal score model.}

\item{Y.form}{Either a character string specifying one formula to be used for all outcome model(s), or a character vector of length 3 specifying formulas for the outcome in compliers in the treatment arm (first element), in noncompliers in the treatment arm (second element), and in controls (third element.).}

\item{Y.model}{The type of model to be used for the outcome. Can be (1) a character string or (2) an object of class family. In case (1), if "linear", "gaussian" or "identity", will fit a linear model; if "binomial", "quasibinomial", "logit", "logistic" or "odds ratio", will fit a quasibinomial model with logit link; if "poisson",  "quasipoisson", "count" or "rate ratio", will fit a quasipoisson model with log link; if "Gamma" or "gamma", will fit a quasi model with log link and variance proportional to mean squared.}

\item{Y.bounds}{A two element vector for the lower and upper bounds of the outcome. Needed when modeling a bounded outcome by a quasibinomial model and/or when conducting a GOR-based sensitivity analysis.}

\item{Y.FUN}{Optional function for customized estimation of outcome mean functions. For an example, see function \code{jobsII_earn3_mus()} in this package, and see how it is used in the vignette. If use \code{Y.FUN}, arguments for outcome model forms and family do not need to be specified, and are ignored.}
}
\description{
Plot the distributions of \mjeqn{\mu_{00}(X)}{ASCII representation} and of \mjeqn{\mu_{01}(X)}{ASCII representation} that are implied by different values of the sensitivity MR parameter. This plot uses side-by-side violin plots.
}
